requirements.txt
Flask==3.0.2
openai==1.12.0
python-dotenv==1.0.1
gunicorn==23.0.0
